+++ rust/gitxetcore/src/utils.rs
@@ -8,8 +8,6 @@ use merklehash::{DataHashHexParseError, MerkleHash};
 use std::collections::HashSet;
 use std::ffi::OsStr;
 use std::path::{Path, PathBuf};
-use std::{io, io::Write};
-use tempfile::NamedTempFile;
 use tracing::error;
 
 /// Find the Oid a ref note references to.
@@ -68,38 +66,6 @@ pub fn shard_path_to_hash(path: &Path) -> Result<MerkleHash, DataHashHexParseErr
     Ok(hash)
 }
 
-/// Write all bytes
-pub fn write_all_file_safe(path: &Path, bytes: &[u8]) -> io::Result<()> {
-    if !path.as_os_str().is_empty() {
-        let dir = path.parent().ok_or_else(|| {
-            io::Error::new(
-                io::ErrorKind::InvalidInput,
-                format!("Unable to find parent path from {path:?}"),
-            )
-        })?;
-
-        // Make sure dir exists.
-        if !dir.exists() {
-            std::fs::create_dir_all(dir)?;
-        }
-
-        let mut tempfile = create_temp_file(dir, "")?;
-        tempfile.write_all(bytes)?;
-        tempfile.persist(path).map_err(|e| e.error)?;
-    }
-
-    Ok(())
-}
-
-pub fn create_temp_file(dir: &Path, suffix: &str) -> io::Result<NamedTempFile> {
-    let tempfile = tempfile::Builder::new()
-        .prefix(&format!("{}.", std::process::id()))
-        .suffix(suffix)
-        .tempfile_in(dir)?;
-
-    Ok(tempfile)
-}
-
 pub fn add_note(repo_path: &Path, notesref: &str, note: &[u8]) -> errors::Result<()> {
     let repo = GitNotesWrapper::open(repo_path, notesref).map_err(|e| {
         error!("add_note: Unable to access git notes at {notesref:?}: {e:?}");
@@ -132,9 +98,7 @@ pub async fn merge_git_notes(base: &Path, head: &Path, notesref: &str) -> errors
 mod test {
     use anyhow::Result;
     use merklehash::*;
-    use std::fs;
     use std::str::FromStr;
-    use tempfile::TempDir;
 
     use crate::utils::*;
 
@@ -157,19 +121,6 @@ mod test {
             assert_eq!(shard_to_meta(&shard), meta);
             assert_eq!(meta_to_shard(&meta), shard);
         }
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_small_file_write() -> Result<()> {
-        let tmp_dir = TempDir::new()?;
-        let bytes = vec![1u8; 1000];
-        let file_name = tmp_dir.path().join("data");
-
-        write_all_file_safe(&file_name, &bytes)?;
-
-        assert_eq!(fs::read(file_name)?, bytes);
 
         Ok(())
     }
