+++ rust/mdb_shard/src/constants.rs
@@ -1,0 +1,4 @@
+pub const MDB_SHARD_TARGET_SIZE: u64 = 64 * 1024 * 1024;
+pub const MDB_SHARD_MIN_TARGET_SIZE: u64 = 48 * 1024 * 1024;
+
+pub const MDB_SHARD_GLOBAL_DEDUP_CHUNK_MODULES: u64 = 1024;
