+++ rust/mdb_shard/src/lib.rs
@@ -1,10 +1,12 @@
 pub mod cas_structs;
+pub mod constants;
 pub mod error;
 pub mod file_structs;
 pub mod intershard_reference_structs;
 pub mod serialization_utils;
 pub mod session_directory;
 pub mod set_operations;
+pub mod shard_dedup_probe;
 pub mod shard_file_handle;
 pub mod shard_file_manager;
 pub mod shard_file_reconstructor;
@@ -13,12 +15,11 @@ pub mod shard_in_memory;
 pub mod shard_version;
 pub mod utils;
 
+pub use constants::MDB_SHARD_TARGET_SIZE;
 pub use intershard_reference_structs::IntershardReferenceSequence;
 pub use shard_file_handle::MDBShardFile;
 pub use shard_file_manager::ShardFileManager;
-pub use shard_format::{
-    MDBShardFileFooter, MDBShardFileHeader, MDBShardInfo, MDB_SHARD_TARGET_SIZE,
-};
+pub use shard_format::{MDBShardFileFooter, MDBShardFileHeader, MDBShardInfo};
 
 // Temporary to transition dependent code to new location
 pub mod shard_file;
