+++ rust/utils/Cargo.toml
@@ -15,6 +15,7 @@ serde = {version = "1.0", features = ["derive"] }
 merklehash = { path = "../merklehash"} 
 xet_error = {path = "../xet_error"}
 futures = "0.3.28"
+tempfile = "3.9.0"
 
 # singleflight
 tokio = {version = "1", features = ["sync"] }
